Renamed SwisClient.py to lowercase and removed samples.
